+++ 05-c#-rest/03-rest-service-actors-supabase/ActorsRestService/Models/Actor.cs
@@ -8,7 +8,7 @@ namespace ActorsRestService.Models
     public partial class Actor: BaseModel
     {
         [PrimaryKey("actor_id")]
-        public int ActorId { get; set; }
+        public long ActorId { get; set; }
 
         [Column("last_name")]
         public string? LastName { get; set; }
